requirements.txt
python-dotenv>=0.15.0
requests>=2.25.1
flask>=1.1.2